+++ src/main/java/com/happiday/Happi_Day/domain/entity/aticle/UserArticleLikes.java
@@ -1,5 +1,6 @@
-package com.happiday.Happi_Day.domain.entity;
+package com.happiday.Happi_Day.domain.entity.aticle;
 
+import com.happiday.Happi_Day.domain.entity.aticle.Article;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -12,8 +13,8 @@ import lombok.experimental.SuperBuilder;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @SuperBuilder(toBuilder = true)
-@Table(name = "user_article_scrap")
-public class UserArticleScrap {
+@Table(name = "user_article_likes")
+public class UserArticleLikes {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
